myproject/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
redis==5.0.4
openai==1.30.0
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.0
pytest-asyncio==0.23.7
PyYAML==6.0.1
